swap the filenames to match their classes
